feat: 이력서 기능 테스트중

+++ src/main/java/org/example/hiro_java/resume/service/dto/ResumesResponse.java
@@ -17,7 +17,7 @@ public class ResumesResponse {
 
     private String resumeId;
     private String applicantName;
-    private Long career;
+    private Integer career;
     private Integer educationLevel;
     private List<String> jobCategories;
     private List<String> languages;
